% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColGen_CuttingStock2.R
\name{ColGen_CuttingStock2}
\alias{ColGen_CuttingStock2}
\title{A Column Generate method in R.}
\usage{
ColGen_CuttingStock2(demand_vec, total_len)
}
\arguments{
\item{demand_vec}{A vector of 2 columns, denotes the needed length and its corresponding number.}

\item{total_len}{the length of unit stock.}
}
\value{
a list containing optimal cost, cutting plan of a stock and the corresponding number needed.
}
\description{
The method is to get continuous solution of cutting stock problem.
}
\examples{
\dontrun{
demand_vec <- matrix(c(3, 6, 7, 25, 20, 18), nrow=3)
total_len = 16
res <- ColGen_CuttingStock2(demand_vec, total_len)
}
}
